Consolidate logging configuration method and tests
br3ndonland/inboard#3

The `configure_logging()` method was previously located in `start.py`,
but the logging configuration dictionary was in `logging_conf.py`. This
organization was not ideal for readability and separation of concerns.
It also created logistical complications, such as requiring the Gunicorn
configuration file `gunicorn_conf.py` to import from start.py, so that
it could configure logging for Gunicorn. The `start.py` script should be
self-contained, and should import other modules and objects from the
package, not the other way around.

This commit will move `configure_logging()` to `logging_conf.py`, and
move the tests to a corresponding module, `test_logging_conf.py`. This
matches up nicely with the latest updates to the `gunicorn_conf.py`, by
having a setup method at the top of the module, and the settings below.
